Moved sort to right class
 - Move sort to right class
-  removed userData.cs
